chore(tag-list): only the one component actually
